Exponential leveling
Remove "kills" feature, implement "experience."

Experience goals are exponential, meaning the
experience goal required for the next level
is always easy to calculate!

Add `Character.next_xp_goal()` method, which
determines the total XP the character needs
for the next level.

`level_up` now returns True/False based on
if the Character leveled up or not.

Change the points to assign in the random
`Character` generator to the level provided.

Change `Character.print_stats()` to display
the experience of the character and its next
XP goal.

Clarity and style corrections.

Decrease number of points assigned when
first creating player, to `4`.

Add miscellaneous notes.